refactor: Recoil 전역 모달 제거후 portal 모달 형성으로 리팩토링
