+++ src/components/common/Modal/ModalPortal.tsx
@@ -1,0 +1,14 @@
+import { ReactNode } from 'react';
+import ReactDom from 'react-dom';
+
+interface Props {
+  children: ReactNode;
+}
+
+const ModalPortal = ({ children }: Props) => {
+  const el = document.getElementById('modal') as HTMLElement;
+
+  return ReactDom.createPortal(children, el);
+};
+
+export default ModalPortal;
