--- src/utils/hooks/useModal.ts
@@ -1,21 +1,0 @@
-import { useRecoilState } from 'recoil';
-import { modalState, ModalType } from 'utils/recoil/ModalRecoil';
-
-export default function useModal() {
-  const [modal, setModal] = useRecoilState(modalState);
-
-  const showModal = ({ modalType, modalProps }: ModalType) => {
-    setModal({ modalType, modalProps });
-  };
-
-  const hideModal = () => {
-    setModal(null);
-  };
-
-  return {
-    modal,
-    setModal,
-    showModal,
-    hideModal,
-  };
-}
